+++ changelog.md
@@ -5,6 +5,12 @@ Tous les changements importants dans l'application seront annotés dans ce fichi
 Le format est basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 et ce projet adhère au [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3] - 2022-02-07
+### Added
+- Ajout Create du CRUD sur Character
+- Ajout Identifier sur Display
+- Ajout colonne Identifier
+
 ## [0.2] - 2022-02-07
 ### Added
 - Ajout Entity Character